bug-fixes
Proxy libraries work right, adding and inserting now as well as descendants for a list of folders you can add a part to. the action menu should always be in view(still a bit glitchy). The doc name in the context menu shouldn't be covered by the X.

bug-fixes

will be cherry picked to bug-fixes branch